@@ -1,14 +1,11 @@
 package ca.pitt.wet.cdts.demo.wetcdtsspringboot;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.BlockJUnit4ClassRunner;
+import org.junit.jupiter.api.Test;
 
-@RunWith(BlockJUnit4ClassRunner.class)
-public class WetCdtsSpringBootApplicationTests {
+class WetCdtsSpringBootApplicationTests {
 
 	@Test
-	public void contextLoads() {
+	void contextLoads() {
 	}
 
 }
