@@ -4,6 +4,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.h2.H2ConsoleAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.context.annotation.Import;
+
+import ca.canada.ised.wet.cdts.WetCdtsSpringBootThymeleafConfig;
 
 /**
  * Application Config.
@@ -12,16 +15,16 @@ import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
  * @since 1.0.0
  */
 @SpringBootApplication(exclude = { DataSourceAutoConfiguration.class, H2ConsoleAutoConfiguration.class })
+@Import(WetCdtsSpringBootThymeleafConfig.class)
 public class ApplicationConfig {
 
-	/**
-	 * Main method.
-	 * 
-	 * @param args
-	 *            <code>String[]</code>
-	 */
-	public static void main(String[] args) {
-		SpringApplication.run(ApplicationConfig.class, args);
-	}
+    /**
+     * Main method.
+     * 
+     * @param args <code>String[]</code>
+     */
+    public static void main(String[] args) {
+        SpringApplication.run(ApplicationConfig.class, args);
+    }
 
 }
