@@ -1,8 +1,9 @@
 package ca.pitt.wet.cdts.demo.wetcdtsspringboot.controllers;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
 /**
@@ -20,7 +21,7 @@ public class LeftNavController {
 	 * 
 	 * @return <code>ModelAndView</code>
 	 */
-	@RequestMapping(method = RequestMethod.GET)
+	@GetMapping
 	public ModelAndView getNav() {
 		return new ModelAndView("nav");
 	}
@@ -30,7 +31,7 @@ public class LeftNavController {
 	 * 
 	 * @return <code>ModelAndView</code>
 	 */
-	@RequestMapping(method = RequestMethod.POST)
+	@PostMapping
 	public ModelAndView postNav() {
 		return new ModelAndView("nav");
 	}
