@@ -1,8 +1,8 @@
 package ca.pitt.wet.cdts.demo.wetcdtsspringboot.controllers.section1;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
 /**
@@ -20,7 +20,7 @@ public class Section1Controller {
      *
      * @return <code>ModelAndView</code>
      */
-    @RequestMapping(path = "/item1", method = RequestMethod.GET)
+    @GetMapping("/item1")
     public ModelAndView getNav() {
         return new ModelAndView("section1/item1");
     }
@@ -30,7 +30,7 @@ public class Section1Controller {
      *
      * @return <code>ModelAndView</code>
      */
-    @RequestMapping(path = "/item2", method = RequestMethod.GET)
+    @GetMapping("/item2")
     public ModelAndView postNav() {
         return new ModelAndView("section1/item2");
     }
